refactor: 리포지토리 인터페이스 추가에 따른 구현체 이름 변경
