+++ src/main/java/lotto/repository/LottoRepository.java
@@ -2,18 +2,10 @@ package lotto.repository;
 
 import lotto.domain.Lotto;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
-public class LottoRepository {
-    private List<Lotto> lottos = new ArrayList<>();
+public interface LottoRepository {
+    public void save(Lotto lotto);
 
-    public void save(Lotto lotto) {
-        lottos.add(lotto);
-    }
-
-    public List<Lotto> findAll() {
-        return lottos;
-    }
+    public List<Lotto> findAll();
 }
